Fixup: Resolve verifier failures from correctness wave
- emit.c: reference inst_t directly before insts is declared
- error.c: use int for npin to match gc_save_len field type
- ns_env.c: add unreachable return after gc_oom_throw for analyzer
- io.c: guard gc_pin/unpin and memcpy against NULL items/buf
- analyze_baseline.txt: regenerate after false-positive delta

+++ src/runtime/error.c
@@ -148,7 +148,7 @@ void set_eval_diag_with_data(mino_state *S, const mino_val *form,
         }
         mino_val *keys[7], *vals[7];
         size_t      n = 0;
-        size_t      npin = 0; /* tracks live gc_pin count for gc_unpin */
+        int         npin = 0; /* tracks live gc_pin count for gc_unpin */
         /* Each mino_keyword/mino_string/mino_int/mino_map call can trigger
          * GC.  Pin every key/val as it is produced so earlier entries are
          * not swept while later ones are being allocated.  All pins are
